requirements.txt
streamlit
google-generativeai
elevenlabs
instagrapi
python-dotenv
requests
Pillow